+++ srcs_otool/parse_argv.c
@@ -18,7 +18,7 @@ static void	read_bin(t_bin *bin, char *source)
 	if (bin->fd > 2)
 	{
 		if (!fstat(bin->fd, &bin->st))
-			bin->ptr = mmap(0, bin->st.st_size, PROT_READ,
+			bin->ptr = mmap(0, (size_t) bin->st.st_size, PROT_READ,
 							MAP_PRIVATE, bin->fd, 0);
 		if (bin->st.st_mode & S_IFDIR)
 		{
